backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
chromadb
google-generativeai
python-multipart
pypdf
python-dotenv
